+++ app/src/components/Cards/common/ArtistAvatar.tsx
@@ -1,0 +1,16 @@
+import { Avatar, AvatarProps } from "@mui/material";
+
+export const ArtistAvatar = ({ ...props }: AvatarProps) => {
+  return (
+    <Avatar
+      slotProps={{
+        img: {
+          referrerPolicy: "no-referrer",
+        },
+      }}
+      alt=""
+      sx={{ width: 42, height: 42 }}
+      {...props}
+    />
+  );
+};