+++ app/src/components/Headers/Track.tsx
@@ -1,8 +1,9 @@
 import { useNavigate } from "react-router-dom";
-import { Avatar, Button, Chip, Stack, Typography } from "@mui/material";
+import { Button, Chip, Stack, Typography } from "@mui/material";
 import { TrackType } from "src/types";
 
 import { DownloadButton } from "../Buttons/DownloadButton";
+import { ArtistAvatar } from "../Cards/common/ArtistAvatar";
 import { ChipQuality } from "../Cards/common/ChipQuality";
 
 import PageHeader from "./Header";
@@ -22,9 +23,8 @@ export default function TrackHeader({ track }: { track: TrackType }) {
       subtitle="Track"
       beforeTitle={
         <Stack direction="row" flexWrap="wrap" spacing={1} alignItems="center">
-          <Avatar
+          <ArtistAvatar
             alt={track.artists?.[0]?.name}
-            sx={{ width: 42, height: 42 }}
             src={`https://resources.tidal.com/images/${track.artists?.[0]?.picture?.replace(
               /-/g,
               "/",